appstories/requirements.txt
moviepy==1.0.3
decorator==4.4.2
audioop-lts>=0.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
proglog==0.1.12
pillow==11.3.0
numpy==2.3.1
ffmpeg-python==0.2.0
pydub==0.25.1
pysrt==1.1.2
python-dotenv==1.1.0
requests==2.32.3
websocket-client==1.8.0
